tests: Move vfio to integration_vfio.rs
Extract mod vfio from integration.rs into a standalone
test binary integration_vfio.rs. This follows the pattern
established by integration_cvm.rs.
